Return packages.json if packages/list.json does not exist
If you go to the drupal 7 repository https://packages.drupal.org/7/ you
will find that it does not support the packages/list.json url.  In that
case we just return the packages.json that was returned.  This was
making retrieving packages.json from a nexus proxy for drupal 7 return a
status 500 error.