fix asset deletion on proxy repositories
Because no hosted facet is attached here asset deletion on zipballs
throws a MissingFacetException when trying to rebuild metadata files
after deleting a zipball.

In proxy repositories package and provider JSON are cached from upstream
and not generated from local repo content, so it's content is still
valid after deletion. Add a conditional and only try to rebuild anything
if a hosted facet is actually attached.